+++ pkg/azuremachinepool/validate_azuremachinepool_create.go
@@ -64,6 +64,11 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return false, microerror.Mask(err)
 	}
 
+	err = checkDataDisks(ctx, azureMPNewCR)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
 	return true, nil
 }