+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Block configuring the `DataDisks` field in AzureMachinePool CRs, and set a default value for it.
+
 ## [1.7.0] - 2020-10-16
 
 ### Added